Move calc_log function
